Change imports to reflect on reorganization
